Don't completely remove compare/set and compare/jmp with constant conditions. Instead, emit equivalent code at final assembly time.
This was the cause of the problem with trashed input-args and
needlessly-preserved saved-regs when a routine had an infinite loop in
it, e.g. (define (x) (while (= 0 0))). By completely removing the
instruction, the live-range-computation code lost track of the
*demand* for the temporary locations of saved regs, and so the
register allocator made unsound decisions for where to put them. With
the new approach, we retain the conditional set/jump, even though we
know it has a fixed outcome, so the register allocator's live-range
computation gets to see it.